[Improve] code style improvement
