+++ streampark-e2e/streampark-e2e-case/src/test/java/org/apache/streampark/e2e/cases/TeamManagementTest.java
@@ -23,6 +23,7 @@ import org.apache.streampark.e2e.core.StreamPark;
 import org.apache.streampark.e2e.pages.LoginPage;
 import org.apache.streampark.e2e.pages.system.SystemPage;
 import org.apache.streampark.e2e.pages.system.TeamManagementPage;
+
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
@@ -34,82 +35,90 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @StreamPark(composeFiles = "docker/basic/docker-compose.yaml")
 public class TeamManagementTest {
-    private static RemoteWebDriver browser;
+  private static RemoteWebDriver browser;
 
-    private static final String userName = "admin";
+  private static final String userName = "admin";
 
-    private static final String password = "streampark";
+  private static final String password = "streampark";
 
-    private static final String teamName = "default";
+  private static final String teamName = "default";
 
-    private static final String newTeamName = "test_new_team";
+  private static final String newTeamName = "test_new_team";
 
-    private static final String newTeamDescription = "test_new_team_description";
+  private static final String newTeamDescription = "test_new_team_description";
 
-    @BeforeAll
-    public static void setup() {
-        new LoginPage(browser)
-            .login(userName, password, teamName)
-            .goToNav(SystemPage.class)
-            .goToTab(TeamManagementPage.class);
-    }
+  @BeforeAll
+  public static void setup() {
+    new LoginPage(browser)
+        .login(userName, password, teamName)
+        .goToNav(SystemPage.class)
+        .goToTab(TeamManagementPage.class);
+  }
 
-    @Test
-    @Order(10)
-    void testCreateTeam() {
-        final TeamManagementPage teamManagementPage = new TeamManagementPage(browser);
-        teamManagementPage.createTeam(newTeamName, newTeamDescription);
+  @Test
+  @Order(10)
+  void testCreateTeam() {
+    final TeamManagementPage teamManagementPage = new TeamManagementPage(browser);
+    teamManagementPage.createTeam(newTeamName, newTeamDescription);
 
-        Awaitility.await().untilAsserted(() -> assertThat(teamManagementPage.teamList())
-            .as("Team list should contain newly-created team")
-            .extracting(WebElement::getText)
-            .anyMatch(it -> it.contains(newTeamName)));
-    }
+    Awaitility.await()
+        .untilAsserted(
+            () ->
+                assertThat(teamManagementPage.teamList())
+                    .as("Team list should contain newly-created team")
+                    .extracting(WebElement::getText)
+                    .anyMatch(it -> it.contains(newTeamName)));
+  }
 
-    @Test
-    @Order(20)
-    void testCreateDuplicateTeam() {
-        final TeamManagementPage teamManagementPage = new TeamManagementPage(browser);
-        teamManagementPage.createTeam(newTeamName, newTeamDescription);
+  @Test
+  @Order(20)
+  void testCreateDuplicateTeam() {
+    final TeamManagementPage teamManagementPage = new TeamManagementPage(browser);
+    teamManagementPage.createTeam(newTeamName, newTeamDescription);
 
-        Awaitility.await().untilAsserted(() -> assertThat(teamManagementPage.errorMessageList())
-            .as("Team Duplicated Error message should be displayed")
-            .extracting(WebElement::getText)
-            .anyMatch(it -> it.contains("Create team failed.")));
+    Awaitility.await()
+        .untilAsserted(
+            () ->
+                assertThat(teamManagementPage.errorMessageList())
+                    .as("Team Duplicated Error message should be displayed")
+                    .extracting(WebElement::getText)
+                    .anyMatch(it -> it.contains("Create team failed.")));
 
-        teamManagementPage.errorMessageConfirmButton().click();
-        teamManagementPage.createTeamForm().buttonCancel().click();
-    }
+    teamManagementPage.errorMessageConfirmButton().click();
+    teamManagementPage.createTeamForm().buttonCancel().click();
+  }
 
-    @Test
-    @Order(30)
-    void testEditTeam() {
-        final TeamManagementPage teamManagementPage = new TeamManagementPage(browser);
-        String editDescription = "edit_" + newTeamDescription;
+  @Test
+  @Order(30)
+  void testEditTeam() {
+    final TeamManagementPage teamManagementPage = new TeamManagementPage(browser);
+    String editDescription = "edit_" + newTeamDescription;
 
-        teamManagementPage.editTeam(newTeamName, editDescription);
+    teamManagementPage.editTeam(newTeamName, editDescription);
 
-        Awaitility.await().untilAsserted(() -> assertThat(teamManagementPage.teamList())
-            .as("Team list should contain edited team")
-            .extracting(WebElement::getText)
-            .anyMatch(it -> it.contains(editDescription)));
-    }
+    Awaitility.await()
+        .untilAsserted(
+            () ->
+                assertThat(teamManagementPage.teamList())
+                    .as("Team list should contain edited team")
+                    .extracting(WebElement::getText)
+                    .anyMatch(it -> it.contains(editDescription)));
+  }
 
-    @Test
-    @Order(40)
-    void testDeleteTeam() {
-        final TeamManagementPage teamManagementPage = new TeamManagementPage(browser);
+  @Test
+  @Order(40)
+  void testDeleteTeam() {
+    final TeamManagementPage teamManagementPage = new TeamManagementPage(browser);
 
-        teamManagementPage.deleteTeam(newTeamName);
+    teamManagementPage.deleteTeam(newTeamName);
 
-        Awaitility.await().untilAsserted(() -> {
-            browser.navigate().refresh();
+    Awaitility.await()
+        .untilAsserted(
+            () -> {
+              browser.navigate().refresh();
 
-            assertThat(
-                teamManagementPage.teamList()
-            ).noneMatch(
-                it -> it.getText().contains(newTeamName)
-            );
-        });
-    }
+              assertThat(teamManagementPage.teamList())
+                  .noneMatch(it -> it.getText().contains(newTeamName));
+            });
+  }
 }
