+++ streampark-e2e/streampark-e2e-case/src/test/java/org/apache/streampark/e2e/pages/LoginPage.java
@@ -19,9 +19,11 @@
 
 package org.apache.streampark.e2e.pages;
 
-import lombok.extern.slf4j.Slf4j;
 import org.apache.streampark.e2e.pages.common.NavBarPage;
 
+import lombok.Getter;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.remote.RemoteWebDriver;
 import org.openqa.selenium.support.FindBy;
@@ -30,79 +32,76 @@ import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-import lombok.Getter;
-import lombok.SneakyThrows;
-
 import java.time.Duration;
 import java.util.List;
 
 @Getter
 @Slf4j
 public final class LoginPage extends NavBarPage {
-    @FindBy(id = "form_item_account")
-    private WebElement inputUsername;
+  @FindBy(id = "form_item_account")
+  private WebElement inputUsername;
 
-    @FindBy(id = "form_item_password")
-    private WebElement inputPassword;
+  @FindBy(id = "form_item_password")
+  private WebElement inputPassword;
 
-    @FindBy(xpath = "//button[contains(@classnames, 'login-button')]")
-    private WebElement buttonLogin;
+  @FindBy(xpath = "//button[contains(@classnames, 'login-button')]")
+  private WebElement buttonLogin;
 
-    private final TeamForm teamForm = new TeamForm();
+  private final TeamForm teamForm = new TeamForm();
 
-    public LoginPage(RemoteWebDriver driver) {
-        super(driver);
+  public LoginPage(RemoteWebDriver driver) {
+    super(driver);
+  }
+
+  @SneakyThrows
+  public NavBarPage login(String username, String password, String teamName) {
+    new WebDriverWait(driver, Duration.ofSeconds(10))
+        .until(ExpectedConditions.elementToBeClickable(buttonLogin));
+
+    inputUsername().sendKeys(username);
+    inputPassword().sendKeys(password);
+    buttonLogin().click();
+
+    try {
+      new WebDriverWait(driver, Duration.ofSeconds(10))
+          .until(ExpectedConditions.visibilityOfAllElements(teamForm.btnSelectTeamDropdown));
+
+      teamForm.btnSelectTeamDropdown.click();
+      teamForm.selectTeam.stream()
+          .filter(it -> it.getText().contains(teamName))
+          .findFirst()
+          .orElseThrow(
+              () -> new RuntimeException(String.format("No %s in team dropdown list", teamName)))
+          .click();
+      teamForm.buttonSubmit.click();
+    } catch (Exception e) {
+      log.warn("No team selection required:", e);
     }
 
-    @SneakyThrows
-    public NavBarPage login(String username, String password, String teamName) {
-        new WebDriverWait(driver, Duration.ofSeconds(10))
-            .until(ExpectedConditions.elementToBeClickable(buttonLogin));
+    new WebDriverWait(driver, Duration.ofSeconds(30))
+        .until(ExpectedConditions.urlContains("/flink/app"));
+    return new NavBarPage(driver);
+  }
 
-        inputUsername().sendKeys(username);
-        inputPassword().sendKeys(password);
-        buttonLogin().click();
-
-        try {
-            new WebDriverWait(driver, Duration.ofSeconds(10))
-                .until(ExpectedConditions.visibilityOfAllElements(teamForm.btnSelectTeamDropdown));
-
-            teamForm.btnSelectTeamDropdown.click();
-            teamForm.selectTeam
-                .stream()
-                .filter(it -> it.getText().contains(teamName))
-                .findFirst()
-                .orElseThrow(() -> new RuntimeException(String.format("No %s in team dropdown list", teamName)))
-                .click();
-            teamForm.buttonSubmit.click();
-        } catch (Exception e) {
-            log.warn("No team selection required:", e);
-        }
-
-        new WebDriverWait(driver, Duration.ofSeconds(30))
-            .until(ExpectedConditions.urlContains("/flink/app"));
-        return new NavBarPage(driver);
+  @Getter
+  public class TeamForm {
+    TeamForm() {
+      PageFactory.initElements(driver, this);
     }
 
-    @Getter
-    public class TeamForm {
-        TeamForm() {
-            PageFactory.initElements(driver, this);
-        }
+    @FindBys({
+      @FindBy(css = "[popupClassName=team-select-popup]"),
+      @FindBy(className = "ant-select-item-option-content")
+    })
+    private List<WebElement> selectTeam;
 
-        @FindBys({
-            @FindBy(css = "[popupClassName=team-select-popup]"),
-            @FindBy(className = "ant-select-item-option-content")
-        })
-        private List<WebElement> selectTeam;
+    @FindBy(css = "[popupClassName=team-select-popup] > .ant-select-selector")
+    private WebElement btnSelectTeamDropdown;
 
-        @FindBy(css = "[popupClassName=team-select-popup] > .ant-select-selector")
-        private WebElement btnSelectTeamDropdown;
+    @FindBy(xpath = "//button[contains(@class, 'ant-btn')]//span[contains(text(), 'OK')]")
+    private WebElement buttonSubmit;
 
-        @FindBy(xpath = "//button[contains(@class, 'ant-btn')]//span[contains(text(), 'OK')]")
-        private WebElement buttonSubmit;
-
-        @FindBy(xpath = "//button[contains(@class, 'ant-btn')]//span[contains(text(), 'Cancel')]")
-        private WebElement buttonCancel;
-    }
+    @FindBy(xpath = "//button[contains(@class, 'ant-btn')]//span[contains(text(), 'Cancel')]")
+    private WebElement buttonCancel;
+  }
 }
