+++ streampark-e2e/streampark-e2e-case/src/test/java/org/apache/streampark/e2e/pages/common/Constants.java
@@ -23,5 +23,5 @@ import lombok.experimental.UtilityClass;
 
 @UtilityClass
 public class Constants {
-    public static final Integer DEFAULT_SLEEP_SECONDS = 1000;
+  public static final Integer DEFAULT_SLEEP_SECONDS = 1000;
 }
