+++ streampark-e2e/streampark-e2e-case/src/test/java/org/apache/streampark/e2e/pages/common/NavBarPage.java
@@ -20,9 +20,10 @@
 
 package org.apache.streampark.e2e.pages.common;
 
-import org.apache.streampark.e2e.pages.apacheflink.ApacheFlinkPage;
+import org.apache.streampark.e2e.pages.flink.ApacheFlinkPage;
 import org.apache.streampark.e2e.pages.system.SystemPage;
 
+import lombok.Getter;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.remote.RemoteWebDriver;
@@ -31,53 +32,62 @@ import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-import lombok.Getter;
-
 import java.time.Duration;
 
 @Getter
 public class NavBarPage {
-    protected final RemoteWebDriver driver;
+  protected final RemoteWebDriver driver;
 
-    @FindBy(xpath = "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'Apache Flink')]")
-    private WebElement apacheFlinkTab;
+  @FindBy(
+      xpath =
+          "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'Apache Flink')]")
+  private WebElement apacheFlinkTab;
 
-    @FindBy(xpath = "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'Resources')]")
-    private WebElement resourcesTab;
+  @FindBy(
+      xpath =
+          "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'Resources')]")
+  private WebElement resourcesTab;
 
-    @FindBy(xpath = "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'Settings')]")
-    private WebElement settingsTab;
+  @FindBy(
+      xpath =
+          "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'Settings')]")
+  private WebElement settingsTab;
 
-    @FindBy(xpath = "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'System')]")
-    private WebElement systemTab;
+  @FindBy(
+      xpath =
+          "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'System')]")
+  private WebElement systemTab;
 
-    public NavBarPage(RemoteWebDriver driver) {
-        this.driver = driver;
-        PageFactory.initElements(driver, this);
+  public NavBarPage(RemoteWebDriver driver) {
+    this.driver = driver;
+    PageFactory.initElements(driver, this);
+  }
+
+  public <T extends NavBarItem> T goToNav(Class<T> nav) {
+    if (nav == ApacheFlinkPage.class) {
+      new WebDriverWait(driver, Duration.ofSeconds(10))
+          .until(ExpectedConditions.elementToBeClickable(apacheFlinkTab));
+      String tabOpenStateXpath =
+          "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'Apache Flink')]/../parent::li[contains(@class, 'streampark-menu-opened')]";
+      if (driver.findElements(By.xpath(tabOpenStateXpath)).isEmpty()) {
+        apacheFlinkTab.click();
+      }
+      return nav.cast(new ApacheFlinkPage(driver));
     }
 
-    public <T extends NavBarItem> T goToNav(Class<T> nav) {
-        if (nav == ApacheFlinkPage.class) {
-            new WebDriverWait(driver, Duration.ofSeconds(10)).until(ExpectedConditions.elementToBeClickable(apacheFlinkTab));
-            String tabOpenStateXpath = "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'Apache Flink')]/../parent::li[contains(@class, 'streampark-menu-opened')]";
-            if (driver.findElements(By.xpath(tabOpenStateXpath)).isEmpty()) {
-                apacheFlinkTab.click();
-            }
-            return nav.cast(new ApacheFlinkPage(driver));
-        }
-
-        if (nav == SystemPage.class) {
-            new WebDriverWait(driver, Duration.ofSeconds(10)).until(ExpectedConditions.elementToBeClickable(systemTab));
-            String tabOpenStateXpath = "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'System')]/../parent::li[contains(@class, 'streampark-menu-opened')]";
-            if (driver.findElements(By.xpath(tabOpenStateXpath)).isEmpty()) {
-                systemTab.click();
-            }
-            return nav.cast(new SystemPage(driver));
-        }
-
-        throw new UnsupportedOperationException("Unknown nav bar");
+    if (nav == SystemPage.class) {
+      new WebDriverWait(driver, Duration.ofSeconds(10))
+          .until(ExpectedConditions.elementToBeClickable(systemTab));
+      String tabOpenStateXpath =
+          "//span[contains(@class, 'ml-2') and contains(@class, 'streampark-simple-menu-sub-title') and contains(text(), 'System')]/../parent::li[contains(@class, 'streampark-menu-opened')]";
+      if (driver.findElements(By.xpath(tabOpenStateXpath)).isEmpty()) {
+        systemTab.click();
+      }
+      return nav.cast(new SystemPage(driver));
     }
 
-    public interface NavBarItem {
-    }
+    throw new UnsupportedOperationException("Unknown nav bar");
+  }
+
+  public interface NavBarItem {}
 }
