+++ streampark-e2e/streampark-e2e-case/src/test/java/org/apache/streampark/e2e/pages/system/entity/UserManagementStatus.java
@@ -20,6 +20,6 @@
 package org.apache.streampark.e2e.pages.system.entity;
 
 public enum UserManagementStatus {
-    LOCKED,
-    EFFECTIVE
+  LOCKED,
+  EFFECTIVE
 }
