+++ streampark-e2e/streampark-e2e-core/src/main/java/org/apache/streampark/e2e/core/Constants.java
@@ -24,18 +24,12 @@ import java.nio.file.Paths;
 
 @UtilityClass
 public final class Constants {
-    /**
-     * tmp directory path
-     */
-    public static final Path HOST_TMP_PATH = Paths.get(System.getProperty("java.io.tmpdir"));
+  /** tmp directory path */
+  public static final Path HOST_TMP_PATH = Paths.get(System.getProperty("java.io.tmpdir"));
 
-    /**
-     * chrome download path in host
-     */
-    public static final Path HOST_CHROME_DOWNLOAD_PATH = HOST_TMP_PATH.resolve("download");
+  /** chrome download path in host */
+  public static final Path HOST_CHROME_DOWNLOAD_PATH = HOST_TMP_PATH.resolve("download");
 
-    /**
-     * chrome download path in selenium/standalone-chrome-debug container
-     */
-    public static final String SELENIUM_CONTAINER_CHROME_DOWNLOAD_PATH = "/home/seluser/Downloads";
+  /** chrome download path in selenium/standalone-chrome-debug container */
+  public static final String SELENIUM_CONTAINER_CHROME_DOWNLOAD_PATH = "/home/seluser/Downloads";
 }
