+++ streampark-e2e/streampark-e2e-core/src/main/java/org/apache/streampark/e2e/core/StreamPark.java
@@ -19,16 +19,16 @@
 
 package org.apache.streampark.e2e.core;
 
+import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
+import org.junit.jupiter.api.TestMethodOrder;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-
-import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
-import org.junit.jupiter.api.TestMethodOrder;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
 @Inherited
 @Testcontainers
@@ -37,5 +37,5 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @TestMethodOrder(OrderAnnotation.class)
 @ExtendWith(StreamParkExtension.class)
 public @interface StreamPark {
-    String[] composeFiles();
+  String[] composeFiles();
 }
